improving upon the integration tests and setting up a library crate
