+++ kernel/Cargo.toml
@@ -20,3 +20,7 @@ features = ["spin_no_std"]
 test-args = ["-device", "isa-debug-exit,iobase=0xf4,iosize=0x04", "-serial", "stdio", "-display", "none"]
 test-success-exit-code = 33
 test-timeout = 180
+
+[[test]]
+name = "should_panic"
+harness = false
