+++ kernel/src/main.rs
@@ -2,12 +2,10 @@
 #![no_std] //disabling std lib.
 #![no_main] //overwriting the entry-point
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
+#![test_runner(operating_system::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
-mod serial;
-mod vga_buffer;
-
+use operating_system::println;
 use core::panic::PanicInfo;
 
 #[cfg(not(test))]
@@ -21,11 +19,7 @@ fn panic(info: &PanicInfo) -> ! {
 #[cfg(test)]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    serial_println!("[FAILED]\n");
-    serial_println!("ERROR: {}\n", info);
-    exit_qemu(QemuExitCode::Failed);
-
-    loop{}
+    operating_system::test_panic_handler(info)
 }
 
 static OSNAME: &[u8] = b"Ziggy OS Loading";
@@ -41,46 +35,6 @@ pub extern "C" fn _start() -> ! { //system entry-point, surprised to learn main 
     loop {}
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failed = 0x11,
-}
-
-pub fn exit_qemu(exit_code: QemuExitCode) {
-    use x86_64::instructions::port::Port;
-
-    unsafe {
-        let mut port = Port::new(0xf4);
-        port.write(exit_code as u32);
-    }
-}
-
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-impl<T> Testable for T
-where
-    T: Fn(),
-{
-    fn run(&self) {
-        serial_print!("{}...\t", core::any::type_name::<T>());
-        self();
-        serial_println!("[ok]");
-    }
-}
-
-#[cfg(test)]
-pub fn test_runner(tests: &[&dyn Fn()]) {
-    serial_println!("Running {} tests", tests.len());
-    for test in tests {
-        test.run();
-    }
-
-    exit_qemu(QemuExitCode::Success);
-}
 
 #[test_case]
 fn trivial_assertion() {
