+++ kernel/tests/should_panic.rs
@@ -1,0 +1,29 @@
+#![no_std]
+#![no_main]
+
+use core::panic::PanicInfo;
+use operating_system::{QemuExitCode, exit_qemu, serial_println, serial_print};
+
+#[panic_handler]
+fn panic(_info: &PanicInfo) -> ! {
+    serial_println!("[ok]");
+    exit_qemu(QemuExitCode::Success);
+
+    loop {}
+}
+
+#[no_mangle]
+pub extern "C" fn _start() -> ! {
+    should_fail();
+    serial_println!("[test did NOT panic]");
+    exit_qemu(QemuExitCode::Failed);
+
+    loop {}
+}
+
+
+#[test_case]
+fn should_fail() {
+    serial_print!("should_panic::should_fail...\t");
+    assert_eq!(2,8);
+}
